ns: move ISystemUpdateControl
