+++ src/core/hle/service/ns/ns.cpp
@@ -26,6 +26,7 @@
 #include "core/hle/service/ns/platform_service_manager.h"
 #include "core/hle/service/ns/read_only_application_control_data_interface.h"
 #include "core/hle/service/ns/read_only_application_record_interface.h"
+#include "core/hle/service/ns/system_update_control.h"
 #include "core/hle/service/ns/vulnerability_manager_interface.h"
 #include "core/hle/service/server_manager.h"
 #include "core/hle/service/set/settings_server.h"
@@ -521,42 +522,6 @@ public:
             {17, nullptr, "PrepareLaunchApplicationFromHost"},
             {18, nullptr, "GetLaunchEvent"},
             {19, nullptr, "GetLaunchResult"},
-        };
-        // clang-format on
-
-        RegisterHandlers(functions);
-    }
-};
-
-class ISystemUpdateControl final : public ServiceFramework<ISystemUpdateControl> {
-public:
-    explicit ISystemUpdateControl(Core::System& system_)
-        : ServiceFramework{system_, "ISystemUpdateControl"} {
-        // clang-format off
-        static const FunctionInfo functions[] = {
-            {0, nullptr, "HasDownloaded"},
-            {1, nullptr, "RequestCheckLatestUpdate"},
-            {2, nullptr, "RequestDownloadLatestUpdate"},
-            {3, nullptr, "GetDownloadProgress"},
-            {4, nullptr, "ApplyDownloadedUpdate"},
-            {5, nullptr, "RequestPrepareCardUpdate"},
-            {6, nullptr, "GetPrepareCardUpdateProgress"},
-            {7, nullptr, "HasPreparedCardUpdate"},
-            {8, nullptr, "ApplyCardUpdate"},
-            {9, nullptr, "GetDownloadedEulaDataSize"},
-            {10, nullptr, "GetDownloadedEulaData"},
-            {11, nullptr, "SetupCardUpdate"},
-            {12, nullptr, "GetPreparedCardUpdateEulaDataSize"},
-            {13, nullptr, "GetPreparedCardUpdateEulaData"},
-            {14, nullptr, "SetupCardUpdateViaSystemUpdater"},
-            {15, nullptr, "HasReceived"},
-            {16, nullptr, "RequestReceiveSystemUpdate"},
-            {17, nullptr, "GetReceiveProgress"},
-            {18, nullptr, "ApplyReceivedUpdate"},
-            {19, nullptr, "GetReceivedEulaDataSize"},
-            {20, nullptr, "GetReceivedEulaData"},
-            {21, nullptr, "SetupToReceiveSystemUpdate"},
-            {22, nullptr, "RequestCheckLatestUpdateIncludesRebootlessUpdate"},
         };
         // clang-format on
